requirements.txt
arabic-reshaper==2.1.0
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
brotlipy==0.7.0
certifi==2020.6.20
cffi @ file:///tmp/build/80754af9/cffi_1605538068321/work
chardet @ file:///tmp/build/80754af9/chardet_1605303185383/work
conda==4.9.2
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1603018141399/work
cryptography @ file:///tmp/build/80754af9/cryptography_1605544487601/work
Django==3.1.1
django-ckeditor==6.0.0
django-etc==1.3.0
django-filter==2.4.0
django-hitcount==1.3.3
django-js-asset==1.2.2
django-widget-tweaks==1.4.8
future==0.18.2
html5lib==1.1
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
isort==5.5.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.3
Pillow==8.1.0
pycodestyle==2.6.0
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
PyMySQL==1.0.2
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1605545627475/work
PyPDF2==1.26.0
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-bidi==0.4.2
pytz==2021.1
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
ruamel-yaml==0.15.87
six @ file:///tmp/build/80754af9/six_1605205327372/work
sqlparse==0.3.1
toml==0.10.1
tqdm @ file:///tmp/build/80754af9/tqdm_1605303662894/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
webencodings==0.5.1
wrapt==1.12.1
xhtml2pdf==0.2.5
